+++ packages/builder/webpack-builder/tests/plugins/svg.test.ts
@@ -1,0 +1,28 @@
+import { expect, describe, it } from 'vitest';
+import { PluginSvg } from '../../src/plugins/svg';
+import { createStubBuilder } from '../../src/shared/stub';
+
+describe('plugins/svg', () => {
+  it('export default url', async () => {
+    const builder = createStubBuilder({
+      plugins: [PluginSvg()],
+    });
+    const config = await builder.unwrapWebpackConfig();
+
+    expect(config).toMatchSnapshot();
+  });
+
+  it('export default Component', async () => {
+    const builder = createStubBuilder({
+      plugins: [PluginSvg()],
+      builderConfig: {
+        output: {
+          svgDefaultExport: 'component',
+        },
+      },
+    });
+    const config = await builder.unwrapWebpackConfig();
+
+    expect(config).toMatchSnapshot();
+  });
+});