+++ packages/runtime/plugin-runtime/src/ssr/serverRender/index.ts
@@ -5,12 +5,6 @@ import SSREntry from './entry';
 import { time } from './measure';
 import { ModernSSRReactComponent } from './type';
 
-declare module '@modern-js/runtime' {
-  interface RuntimeContext {
-    ssrContext?: any;
-  }
-}
-
 export const render = async (
   ctx: RuntimeContext,
   _: string = process.cwd(),
@@ -18,9 +12,9 @@ export const render = async (
 ): Promise<string> => {
   const { ssrContext } = ctx;
 
-  return run(ssrContext.request.headers, async () => {
+  return run(ssrContext!.request.headers, async () => {
     const entry = new SSREntry({
-      ctx: ssrContext,
+      ctx: ssrContext!,
       App,
     });
     entry.metrics.emitCounter('app.visit.count', 1);
@@ -34,7 +28,7 @@ export const render = async (
 
     const cacheConfig = PreRender.config();
     if (cacheConfig) {
-      ctx.ssrContext.cacheConfig = cacheConfig;
+      ctx.ssrContext!.cacheConfig = cacheConfig;
     }
 
     return html;