+++ packages/cli/babel-preset-base/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/babel-preset-base
 
+## 1.16.0
+
+### Patch Changes
+
+- Updated dependencies [641592f52]
+- Updated dependencies [3904b30a5]
+- Updated dependencies [1100dd58c]
+- Updated dependencies [e04e6e76a]
+- Updated dependencies [81c66e4a4]
+- Updated dependencies [2c305b6f5]
+  - @modern-js/utils@1.16.0
+
 ## 1.15.0
 
 ### Patch Changes
@@ -15,7 +27,7 @@
 
 ### Minor Changes
 
-- 7b9067f: add babel plugin for web builder
+- 7b9067f: add babel plugin for webpack-builder
 
 ### Patch Changes