+++ packages/runtime/plugin-garfish/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @modern-js/plugin-garfish
 
+## 1.16.1
+
+### Patch Changes
+
+- fix: replaceAll not work in node14
+
+## 1.16.0
+
+### Minor Changes
+
+- 1100dd58c: chore: support react 18
+
+  chore: 支持 React 18
+
+### Patch Changes
+
+- Updated dependencies [641592f52]
+- Updated dependencies [3904b30a5]
+- Updated dependencies [1100dd58c]
+- Updated dependencies [a480d6ad0]
+- Updated dependencies [e04e6e76a]
+- Updated dependencies [81c66e4a4]
+- Updated dependencies [2c305b6f5]
+  - @modern-js/utils@1.16.0
+  - @modern-js/runtime@1.16.0
+
 ## 1.15.0
 
 ### Patch Changes