+++ packages/runtime/plugin-runtime/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @modern-js/runtime
 
+## 1.16.1
+
+### Patch Changes
+
+- fix: runtime type
+
+  fix: 修复 runtime 类型定义
+
+## 1.16.0
+
+### Minor Changes
+
+- 1100dd58c: chore: support react 18
+
+  chore: 支持 React 18
+
+### Patch Changes
+
+- a480d6ad0: fix: remove helmet regexp's global tag
+  fix: 删除 helmet 中正则匹配的全局标记
+- Updated dependencies [641592f52]
+- Updated dependencies [3904b30a5]
+- Updated dependencies [1100dd58c]
+- Updated dependencies [e04e6e76a]
+- Updated dependencies [81c66e4a4]
+- Updated dependencies [2c305b6f5]
+  - @modern-js/utils@1.16.0
+  - @modern-js/webpack@1.16.0
+  - @modern-js/plugin@1.16.0
+
 ## 1.15.0
 
 ### Patch Changes